Move elastix to separate module
