+++ docs/api/registration.md
@@ -2,4 +2,8 @@
     options:
       members:
         - register
+
+::: multiview_stitcher.elastix
+    options:
+      members:
         - registration_ITKElastix
