+++ src/multiview_stitcher/browser/specs.py
@@ -17,10 +17,10 @@ from multiview_stitcher import fusion, registration
 from multiview_stitcher.browser import elastix
 
 #: Pairwise registration functions selectable from the browser. `itk_elastix`
-#: is elastix compiled to WebAssembly - see `browser.elastix` - rather than
-#: `registration.registration_ITKElastix`, whose native extension has no
-#: WebAssembly build. Naming it here costs nothing: the module imports
-#: itkwasm-elastix only once a registration actually asks for it.
+#: is `elastix.registration_ITKElastix` on the WebAssembly backend - the only
+#: elastix that exists in the browser, see `browser.elastix`. Naming it here
+#: costs nothing: itkwasm-elastix is imported only once a registration
+#: actually asks for it.
 PAIRWISE_REGISTRATION_FUNCS = {
     "phase_correlation": registration.phase_correlation_registration,
     "marker_based": registration.registration_marker_based,
